MELT: Factorize implementation of syntehtic reasoning scenarios
